+++ contracts/adapters/interfaces/IGuildKick.sol
@@ -35,11 +35,11 @@ interface IGuildKick {
         bytes calldata data
     ) external returns (uint64);
 
-    function guildKick(DaoRegistry dao, uint64 proposalId) external;
+    function guildKick(DaoRegistry dao, uint256 proposalId) external;
 
     function rageKick(
         DaoRegistry dao,
-        uint64 proposalId,
+        uint256 proposalId,
         uint256 toIndex
     ) external;
 }
